Simplify the handling of escaped delimiters.
Position 5 of the list returned by syntax-ppss is non-nil iff
the character is escaped according to the syntax table. Use it to
remove the special case for Lisps other than Emacs Lisp.

The special case for Emacs Lisp is still there, to catch ?(.
This form, however, is deprecated, confuses many a minor mode and
confuses the emacs-lisp-mode indentation engine, so the special case
will likely go away soon, too.