Remove the Emacs Lisp specific escaping test.
In fact, nuke the infrastructure altogether.

Parens as characters without a backslash are nasty and they really should
never be used:
 - This syntax is discouraged by the GNU Emacs Lisp Reference Manual.
 - They confuse many a minor mode, such as show-paren-mode.
 - The way we currently handle them partially solves the problem, but
   is a hack made of eldritch magic, implemented by gluing special cases
   together hoping they'll work...
 - ...but it doesn't really work, and it's pretty easy to get
   mishighlighted delimiters anyway.
 - They confuse the Emacs Lisp indentation engine. Really.
 - It's the major mode's job to set the syntax-table properties on
   individual characters.

Because of these reasons, it's just easier to drop this misfeature
and rely on people to start using ?\( instead.

And because the whole point of
rainbow-delimiters-escaped-char-predicate-list is to allow for this,
nuke it too. It was never *really* a documented part of the API.